requirements.txt
google-adk>=0.1.0
google-generativeai>=0.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
serpapi>=0.1.0
python-dotenv>=1.0.0
certifi>=2024.2.2
SpeechRecognition>=3.10.0
pyttsx3>=2.90
PyAudio>=0.2.13  # For microphone access
# Add other dependencies as needed, e.g., for specific voice capabilities later
typer>=0.9.0
selenium>=4.15.0
opencv-python>=4.8.0
pyautogui>=0.9.54
pytesseract>=0.3.10
pillow>=10.0.0
numpy>=1.24.0  # Required by OpenCV
schedule>=1.2.0  # For task scheduling
pywin32>=306; platform_system=="Windows"  # For Windows application control
psutil>=5.9.0  # For system information
pycaw>=20240210; platform_system=="Windows"  # For volume control
wmi>=1.5.1; platform_system=="Windows"  # For screen brightness control 